+++ src/_modules/root/root.controller.ts
@@ -4,14 +4,11 @@ import consts from '~config/consts';
 class RootController {
   constructor(){}
 
-  async helloWorld({set, user, session, log}: any) {
-    log.error(user, "User");
-    log.info(session, "Session");
-
+  async helloWorld({set, user, session}: any) {
     try {
-      console.log(user);
-      console.log(session);
-      console.log(user?.profileId ?? null);
+      console.log("User: ",user);
+      console.log("Session: ",session);
+      console.log("ProfileID: ",user?.profileId ?? null);
       
       return { message: `Welcome to the ${consts.server.name} Server! Version ${consts.server.version}` }
     } catch (error) {
@@ -24,8 +21,9 @@ class RootController {
 
   async helloTime({ user, session, params, store:{timezone} }: any) {
 
-    console.warn(user);
-    console.warn(session);
+    console.log("User: ",user);
+    console.log("Session: ",session);
+    console.log("ProfileID: ",user?.profileId ?? null);
 
     const currentHour = new Date().getHours();
 
@@ -42,7 +40,7 @@ class RootController {
     }
 
     try {
-      return { message: `It's a beautiful ${getGreeting()} in ${timezone ?? 'your area'} isn't it ${params?.name || ''}.` }
+      return { message: `It's a beautiful ${getGreeting()} in ${timezone ?? 'your area'} isn't it ${params?.name || ''}?` }
     } catch (error) {
       console.error(error);