+++ src/_modules/users/users.handler.ts
@@ -1,21 +1,27 @@
 import Elysia, { t } from "elysia";
 import { UsersController } from ".";
-import { checkCookieAuth, checkForProfile, checkIsAdmin, checkIsStaff } from "~middleware/Auth";
-import { UserQueriesDTO,AutoUserBodyDTO, AutoUserResponseDTO, UserResponseDTO, ProfileQueriesDTO, ProfileResponseDTO, ProfileBodyDTO, UpdateProfileDTO } from "./users.model";
+import { checkAuth, checkCookieAuth, checkEmailVerified, checkForProfile, checkIsAdmin, checkIsStaff } from "~middleware/authChecks";
+import { UserQueriesDTO,AutoUserBodyDTO, AutoUserResponseDTO, UserResponseDTO, ProfileQueriesDTO, ProfileResponseDTO, ProfileBodyDTO, UpdateProfileBodyDTO } from "./users.model";
 
 const usersHandler = new Elysia({
     prefix: '/users',
     detail: { description:'User management endpoint', tags: ['Users'] }
 })
     // Lifecycle, auth
-    .onBeforeHandle(checkCookieAuth)
+    .onBeforeHandle([checkAuth])
 
     // Get all Users [STAFF]
     .get('/', UsersController.getAllUsers, {
         beforeHandle: [ checkForProfile, checkIsStaff ],
-        query: t.Object({ isActive: t.Optional(t.Boolean()) }),
+        query: t.Object({
+            isActive: t.Optional(t.BooleanString()),
+            profiles: t.Optional(t.BooleanString()),
+        }),
         response: {
-            200: t.Undefined( t.Object({ data: t.Array(UserResponseDTO), message: t.Optional( t.String({ default: 'Successfully retrieved Users' }) ) }) ),
+            200: t.Union([
+                t.Object({ data: t.Array(UserResponseDTO), message: t.Optional( t.String({ default: 'Successfully retrieved Users' }) ) }),
+                t.Undefined()
+            ]),
             404: t.Object({ message: t.String({ default: 'No Users found' }) }),
             500: t.Object({ message: t.String({ default: 'Could not fetch Users' }) })
         }
@@ -32,24 +38,31 @@ const usersHandler = new Elysia({
     })
 
     // Get single User by ID [STAFF]
-    .get('/user/:id', UsersController.getAccountById, {
+    .get('/user/:userId', UsersController.getAccountById, {
+        beforeHandle: [checkIsStaff],
         params: t.Object({
-            id: t.String()
+            userId: t.String()
         }),
         query: UserQueriesDTO,
         response: { 
-            200: t.Undefined(t.Object({ data: UserResponseDTO, message: t.String({ default: 'Successfully retrieved User' }) }) ),
+            200: t.Union([
+                t.Object({ data: UserResponseDTO, message: t.String({ default: 'Successfully retrieved User' }) }),
+                t.Undefined()
+            ]),
             404: t.Object({ message: t.String({ default: 'User with that ID not found' }) }),
             500: t.Object({ message: t.String({ default: 'Could not fetch a User' }) })
         }
     })
 
-    .get('/profile/:id', UsersController.getProfileById,{
-        beforeHandle: [],
-        params: t.Object({ id: t.String() }),
+    .get('/profile/:id', UsersController.getProfileByUserId,{
+        beforeHandle: [checkIsStaff],
+        params: t.Object({ userId: t.String() }),
         query: ProfileQueriesDTO,
         response: {
-            200: t.Undefined(t.Object({ data: ProfileResponseDTO, message: t.String({ default:'Successfully retrieved User Profile' }) })),
+            200: t.Union([
+                t.Object({ data: ProfileResponseDTO, message: t.String({ default:'Successfully retrieved User Profile' }) }),
+                t.Undefined()
+            ]),
             404: t.Object({ message: t.String({ default:'Profile not available' })}),
             500: t.Object({ message: t.String({ default:'Could not load User Profile of that ID' })}),
         }
@@ -59,17 +72,20 @@ const usersHandler = new Elysia({
         query: ProfileQueriesDTO,
         response: {
             200: t.Object({ data: ProfileResponseDTO, message: t.String({ default:'Successfully retrieved your User Profile' }) }),
-            403: t.Object({ message: t.String({ default:'Profile is deactivated' })}),
             404: t.Object({ message: t.String({ default:'Profile does not exist' })}),
+            406: t.Object({ message: t.String({ default:'Profile is deactivated' })}),
             500: t.Object({ message: t.String({ default:'Could not load User Profile of that ID' })}),
         },
     })
 
     .get('/profiles', UsersController.getAllProfiles, {
-        beforeHandle: [],
+        beforeHandle: [checkIsStaff],
         query: ProfileQueriesDTO,
         response: {
-            200: t.Undefined(t.Object({ profiles: t.Array(ProfileResponseDTO), message: t.String({ default: 'No User Profiles found' }) })),
+            200: t.Union([
+                t.Object({ data: t.Array(ProfileResponseDTO), message: t.String({ default: 'Successfully retrieved n User Profiles' }) }),
+                t.Undefined()
+            ]),
             404: t.Object({message: t.String({ default: 'No User Profiles found' }) }),
             500: t.Object({message: t.String({ default: 'Unable to fetch User Profiles' }) })
         },
@@ -78,9 +94,12 @@ const usersHandler = new Elysia({
 
     // Add a new Post User Account [ADMIN]
     .get('/autousers', UsersController.getAllAutoEnrollers,{
-        beforeHandle:[checkIsAdmin],
+        beforeHandle:[checkIsStaff],
         response: {
-            200: t.Undefined(t.Object({ data: t.Array(AutoUserResponseDTO), message: t.String({ default: 'Retrieved all Auto-Users' }) }) ),
+            200: t.Union([
+                t.Object({ data: t.Array(AutoUserResponseDTO), message: t.String({ default: 'Retrieved all Auto-Users' }) }),
+                t.Undefined()
+            ]),
             500: t.Object({ message: t.String({ default: 'Could not fetch Auto-Users.'}) })
         }
     })
@@ -91,10 +110,15 @@ const usersHandler = new Elysia({
 
     // Create new User Profile [SELF]
     .post('/profile', UsersController.createNewProfile,{
+        beforeHandle: [checkEmailVerified],
         body: ProfileBodyDTO,
         response: {
-            201: t.Object({ data: ProfileResponseDTO, message: t.String({ default: 'Successfullly created new User Profile' }) }),
+            201: t.Union([
+                t.Object({ data: ProfileResponseDTO, message: t.String({ default: 'Successfully created new User Profile' }) }),
+                t.Undefined()
+            ]),
             302: t.Object({ message: t.String({ default: 'A profile already exists with those credentials'}) }),
+            403: t.Object({ message: t.String({ default: 'Your account is not verified.'}) }),
             406: t.Object({ message: t.String({ default: 'Your submission was not valid.'}) }),
             409: t.Object({ message: t.String({ default: 'You already have a profile'}) }),
             500: t.Object({ message: t.String({ default: 'Problem processing profile submission.'}) })
@@ -106,7 +130,10 @@ const usersHandler = new Elysia({
         beforeHandle:[checkIsAdmin],
         body: AutoUserBodyDTO,
         response: {
-            201: t.Undefined(t.Object({ data: AutoUserResponseDTO, message: t.String({ default: 'Successfullly addedd an Auto-User' }) }) ),
+            201: t.Union([
+                t.Object({ data: AutoUserResponseDTO, message: t.String({ default: 'Successfullly addedd an Auto-User' }) }),
+                t.Undefined()
+            ]),
             500: t.Object({ message: t.String({ default: 'Could not create Auto-User.'}) })
         }
     })
@@ -118,38 +145,43 @@ const usersHandler = new Elysia({
     // Update User Profile [SELF]
     .put('/profile', UsersController.updateUserProfile, {
         query: ProfileQueriesDTO,
-        body: UpdateProfileDTO,
+        body: UpdateProfileBodyDTO,
         response: {
             200: t.Object({ data: ProfileResponseDTO, message: t.String({ default: 'Successfully updated User Profile.'}) }),
-            404: t.Object({ message: t.String({ default: 'Profile not found' })}),
+            404: t.Object({ message: t.String({ default: 'Could not update Profile' })}),
             500: t.Object({ message: t.String({ default: 'Could not modify User Profile.' })})
         }
     })
 
-    // Update User Profile [STAFF]
+    // Update other User Profile [STAFF]
     .put('/profile/:userId', UsersController.updateUserProfile, {
+        beforeHandle: [checkIsStaff],
         params: t.Object({ userId: t.String() }),
         query: ProfileQueriesDTO,
-        body: UpdateProfileDTO,
+        body: UpdateProfileBodyDTO,
         response: {
-            200: t.Object({ data: ProfileResponseDTO, message: t.String({ default: 'Successfully updated User Profile.'}) }),
-            404: t.Object({ message: t.String({ default: 'Profile not found' })}),
+            200: t.Union([
+                t.Object({ data: ProfileResponseDTO, message: t.String({ default: 'Successfully updated User Profile.'}) }),
+                t.Undefined()
+            ]),
+            404: t.Object({ message: t.String({ default: 'Could not update Profile' })}),
             500: t.Object({ message: t.String({ default: 'Could not modify User Profile.' })})
         }
     })
 
     // Activate/Deactivate User Profile [STAFF]
-    .put('/deactivate/user/:id', UsersController.deactivateUser, {
-        params: t.Object({ userid: t.String() }),
-        body: t.Object({ isActive: t.Boolean() }),
+    .put('/deactivate/user/:userId', UsersController.deactivateUser, {
+        beforeHandle: [checkIsAdmin],
+        params: t.Object({ userId: t.String() }),
+        body: t.Object({ isComment: t.String() }),
         response: {
-            200: t.Object({ data: t.Optional(UserResponseDTO), message: t.String() }),
+            200: t.Union([
+                t.Object({ data: t.Optional(UserResponseDTO), message: t.String() }),
+                t.Undefined()
+            ])
             // 404: t.Object({ message: t.String({ default: 'Profile not found' })}),
             // 500: t.Object({ message: t.String({ default: 'Could not modify User Profile.' })})
         }
     })
-
-
-    
 
 export default usersHandler;